bin/heatmap.py(load_file): "Fix" warning
"FutureWarning: The behavior of DataFrame concatenation with empty or
all-NA entries is deprecated. In a future version, this will no longer
exclude empty or all-NA columns when determining the result dtypes. To
retain the old behavior, exclude the relevant entries before the concat
operation."

Using `np.nan` instead of pd.NA is kinda a cheeky way to dodge this, but
I'll take it >u<

* bin/heatmap.py(load_file): "Fix" warning