analysis: Add GenerationStrategy::DDAKOSimulatedAnnealing
I removed several "features" from the version haphazardly developed
outside of version control, including. a closure that prints
throughput info on every N invocations, dynamic resets w/ heat
adjustments, and the multiple-profile system that I use for
cross-scenario optimization.

I left all that out mostly because the code was ugly, or (in the case
of dynamic reset) because I didn't find them useful in practice (Freya
called that one). I did retain some light heat adjustments that
evolved out of a "plateau" system, broadly clamping temp down early on
and letting up as/if scores improve, getting to the meat faster and
spending more time on successful lineages. This seems to work well for
me, but might not generalize.