ddako_sim_annl: Add feedback TUI
I mentioned in my commit message that I'd omitted a rate tracker.
Well, it just wasn't as much run without it.
I'm using `atty` to disable it when `STDOUT` isn't a TTY.
And I'm not enabling it for other strats yet, until/unless I see a
reason to -- something like Keymui will be the future anyway.